+++ Shared.Contracts/Events/StockUpdateEvent.cs
@@ -3,5 +3,5 @@ namespace Shared.Contracts.Events;
 // Stores response to the InventorySyncEvent
 public class StockUpdateEvent : IStoreEvent
 {
-    public int EnterpriseId { get; set; }
+    public long EnterpriseId { get; set; }
 }
